Revert version change of Astor. Rename test.
