+++ src/maven-repair/pom.xml
@@ -92,7 +92,7 @@
         <dependency>
             <groupId>org.inria.sacha.automaticRepair</groupId>
             <artifactId>astor</artifactId>
-            <version>20190919</version>
+            <version>0.0.2-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>fr.inria.gforge.spirals</groupId>
